Now more data in TableDatabaseMultiKeys
Now TableDatabaseMultiKeys can save more data (pricipal ID is Long)

+++ src/me/pepe/DatabaseAPI/DatabaseManager/Types/TableDatabaseMultiKeys.java
@@ -4,12 +4,11 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import me.pepe.DatabaseAPI.DatabaseManager.DatabaseKeyType;
-import me.pepe.DatabaseAPI.DatabaseManager.Databases.ServerDataDatabase;
 import me.pepe.DatabaseAPI.DatabaseManager.Tables.DatabaseTable;
 
 public abstract class TableDatabaseMultiKeys extends DatabaseTable {
 	public TableDatabaseMultiKeys(String name, String keyName, Database database) {
-		super(name, keyName, DatabaseKeyType.INT, database);
+		super(name, keyName, DatabaseKeyType.LONG, database);
 		hasData = true;
 		setAutoIncrement(true);
 	}
